Refactor: Move response to utils and rename dependencies to use_cases
