+++ src/use_cases/qr_deps.py
