+++ src/use_cases/user_deps.py
@@ -6,9 +6,9 @@ from email.mime.text import MIMEText
 from passlib.context import CryptContext
 
 from src.config import email_config, settings, secrets
-from src.dependencies.qr_deps import generate_key_qr
-from src.dependencies.token_deps import generate_token_jwt
 from src.models.user import User, StoredUser
+from src.use_cases.qr_deps import generate_key_qr
+from src.use_cases.token_deps import generate_token_jwt
 
 context = CryptContext(
     schemes=[secrets.get("CRYPTOCONTEXT_SCHEM")],
