+++ src/utils/__init__.py
@@ -1,5 +1,5 @@
 from .date_time import DateTime
-from .responses import UserMessage, LoginMessage
+from .response_message import UserMessage, LoginMessage
 
 __all__ = [
     'UserMessage',
